backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
firebase-admin==6.4.0
google-cloud-firestore==2.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
httpx==0.25.2
openai==1.3.7
anthropic==0.7.8
requests==2.31.0
cors==1.0.1